// src/enums/DayPhase.ts
enum DayPhase {
  NIGHT = 'NIGHT',
  NIGHT_START = 'NIGHT_START',
  NIGHT_END = 'NIGHT_END',
  NAUTICAL_DAWN = 'NAUTICAL_DAWN',
  DAWN = 'DAWN',
  SUNRISE = 'SUNRISE',
  SUNRISE_GOLDEN_HOUR = 'SUNRISE_GOLDEN_HOUR',
  ZENITH = 'ZENITH',
  DAY = 'DAY',
  SUNSET_GOLDEN_HOUR = 'SUNSET_GOLDEN_HOUR',
  SUNSET = 'SUNSET',
  DUSK = 'DUSK',
  NAUTICAL_DUSK = 'NAUTICAL_DUSK',
  NADIR = 'NADIR',
}
export default DayPhase